pyproject.toml
[project]
name = "DataOrchestra"
version = "0.1.0"
description = "A toolkit for cleaning and preparing text datasets for LLM training and finetuning"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Pratyay Mustafi", email = "pratyaymustafi@outlook.com"}
]
keywords = ["machine-learning", "data-cleaning", "nlp", "llm", "text-processing", "dataset-preparation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General",
]
requires-python = ">=3.14"

dependencies = [
    "beautifulsoup4>=4.12.0",
    "python-docx>=0.8.11",
    "duckduckgo_search>=7.0.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",
    "Pillow>=8.0.0",
    "requests>=2.25.0",
    "tqdm>=4.62.0",
    # New dependencies for refactored text processing
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "langdetect>=1.0.9",
    "aiofiles>=23.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tenacity>=8.2.0",
    "cachetools>=5.3.0",
    "aiohttp>=3.9.0",
    "pytest>=9.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    "hypothesis>=6.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.5.0",
    "twine>=4.0.0",
    "build>=0.10.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]
all = [
    "beautifulsoup4>=4.12.0",
    "python-docx>=0.8.11",
    "ddgs>=0.2.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",
    "Pillow>=8.0.0",
    "requests>=2.25.0",
    "tqdm>=4.62.0",
]

[project.urls]
Homepage = "https://github.com/your-username/DataOrchestra"
Repository = "https://github.com/your-username/DataOrchestra.git"
Documentation = "https://dataorchestra.readthedocs.io"
"Bug Tracker" = "https://github.com/your-username/DataOrchestra/issues"


[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["DataOrchestra*"]
exclude = ["backup*"]

[tool.setuptools.package-data]
DataOrchestra = ["py.typed"]
